+++ test/dredd_hook.js
@@ -17,34 +17,6 @@ hooks.before("Public Space 🏡 > Prague City Districts > GET District", (transa
     transaction.fullPath = transaction.fullPath.replace("praha-1", storage["districtSlug"]);
 });
 
-hooks.before(
-    "Traffic 🚘 > Shared Cars > GET All Shared Cars",
-    (transaction) => (transaction.skip = true)
-);
-
-hooks.before(
-    "Traffic 🚘 > Shared Cars > GET Shared Car",
-    (transaction) => (transaction.skip = true)
-);
-
-hooks.after("Traffic 🚘 > Bicycle Parkings > GET All Bicycle Parkings", (transaction) => {
-    storage["bicycleParkingId"] = JSON.parse(transaction.real.body).features[0].properties.id;
-});
-
-hooks.before("Traffic 🚘 > Bicycle Parkings > GET Bicycle Parking", (transaction) => {
-    transaction.request.uri = transaction.request.uri.replace("282743091", storage["bicycleParkingId"]);
-    transaction.fullPath = transaction.fullPath.replace("282743091", storage["bicycleParkingId"]);
-});
-
-hooks.after("Traffic 🚘 > Traffic Cameras > GET All Traffic Cameras", (transaction) => {
-    storage["id"] = JSON.parse(transaction.real.body).features[0].properties.id;
-});
-
-hooks.before("Traffic 🚘 > Traffic Cameras > GET Traffic Camera", (transaction) => {
-    transaction.request.uri = transaction.request.uri.replace("404032", storage["id"]);
-    transaction.fullPath = transaction.fullPath.replace("404032", storage["id"]);
-});
-
 hooks.before("Public Space 🏡 > Gardens > GET All Gardens", (transaction) => (transaction.skip = true));
 
 hooks.before("Public Space 🏡 > Gardens > GET Garden", (transaction) => (transaction.skip = true));
@@ -65,15 +37,6 @@ hooks.after("Public Space 🏡 > Playgrounds > GET All Playgrounds", (transactio
 });
 
 hooks.before("Public Space 🏡 > Playgrounds > GET Playground", (transaction) => {
-    transaction.request.uri = transaction.request.uri.replace("72", storage["id"]);
-    transaction.fullPath = transaction.fullPath.replace("72", storage["id"]);
-});
-
-hooks.after("Public Space 🏡 > Public Toilets > GET All Public Toilets", (transaction) => {
-    storage["id"] = JSON.parse(transaction.real.body).features[0].properties.id;
-});
-
-hooks.before("Public Space 🏡 > Public Toilets > GET Public Toilet", (transaction) => {
     transaction.request.uri = transaction.request.uri.replace("72", storage["id"]);
     transaction.fullPath = transaction.fullPath.replace("72", storage["id"]);
 });