+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [3.5.0] - 2023-04-12
+
+### Changed
+
+-   removed modules: bicycle-parkings, parking-zones, public-toilets, shared-cars, meteosensors, traffic-cameras
+-   removed general-routes
+
 ## [3.4.7] - 2023-04-03
 
 -   No changelog